Move reference cleanup logic to its own class
This isolates and encapsulates the cleanup logic, so the
JythonScriptLanguage itself remains easy to understand.
